todo project backend complete  and user register and login logout complete

+++ controllers/taskController.js
@@ -3,9 +3,9 @@ const client = require('../config/redisConfig')
 
 const createTask = async(req,res) => {
     const {title,description} = req.body
-
+    const userId = req.session.user?.userId || req.userId
     try {
-        const addTask = new Task({title,description})
+        const addTask = new Task({title,description,user : userId})
         await addTask.save()
 
         return res.status(201).json({
@@ -22,6 +22,7 @@ const createTask = async(req,res) => {
 
 const getTask = async (req, res) => {
   try {
+    const userId = req.session.user?.userId || req.userId
     const cacheValue = await client.get('todos');
 
     if (cacheValue) {
@@ -33,7 +34,7 @@ const getTask = async (req, res) => {
       });
     }
 
-    const getTodo = await Task.find({});
+    const getTodo = await Task.find({userId});
 
     if (!getTodo || getTodo.length === 0) {
       return res.status(200).json({ success: true, todo: [] });
@@ -56,9 +57,16 @@ const getTask = async (req, res) => {
 
 const deleteTask = async(req,res) => {
     const {id} = req.params
-
+   const userId = req.session.user?.userId || req.userId
     try {
-        const deleteTask = await Task.findByIdAndDelete(id)
+        const deleteTask = await Task.findByIdAndDelete({id,user:userId})
+
+        if(!deleteTask){
+          return res.status(404).json({
+            success:false,
+            message: 'todo not found'
+          })
+        }
         return res.status(200).json({
             success:true,
             message:'deleted successfULLY'
@@ -72,10 +80,15 @@ const deleteTask = async(req,res) => {
 
 const updateTask = async(req,res) => {
     const {id} = req.params
-    const {title,description,complete} = req.body
+    const userId = req.session.user?.userId || req.userId
     try {
-        const update = await Task.findByIdAndUpdate(id,{title,description,complete},{new:true})
-
+        const update = await Task.findByIdAndUpdate({id,user : userId},{$set: req.body},{new:true,runValidators : true})
+          if(!update){
+            return res.status(404).json({
+              success:false,
+              message:'Todo not found'
+            })
+          }
         return res.status(200).json({
             success:true,
             message:'update successFully'
@@ -86,5 +99,32 @@ const updateTask = async(req,res) => {
         })
     }
 }
+ const completeTask = async (req,res) => {
+  try {
+    const userId = req.session.user?.userId || req.userId
+    const {id} = req.params
+    const todo = await Task.findOne({id,user:userId})
 
-module.exports = {createTask,getTask,deleteTask,updateTask}
+    if(!todo){
+      return res.json({
+        success:false,
+        message:"todo not found"
+      })
+    }
+
+    todo.complete = !todo.complete
+    await todo.save()
+
+    res.json({
+      success:true,
+      message: `Todo marked as ${todo.complete ? 'complete' :'incomplete'}`,
+      todo
+    })
+  } catch (error) {
+    res.status(500).json({
+      success:false,
+      message:'error updating todo'
+    })
+  }
+ }
+module.exports = {createTask,getTask,deleteTask,updateTask,completeTask}
